api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
